config replacement respects IP address changes, pings out after restore
Modified the configuration file copy routine to replace the IP address
of the backup config file with the IP address assigned to the container
before restoration. Also replaces the IP address stored within each
snapshot configuration with the current IP address.

Enabled the function to ping google.com from the within the container in
order to force an arp update if necessary.